refactor(graph): isolate bn254 evaluator
